+++ jscomp/frontend/ast_derive_projector.ml
@@ -55,7 +55,7 @@ let init () =
                     {
                       pcd_name = {loc; txt = con_name};
                       pcd_args;
-                      pcd_loc = _;
+                      pcd_loc;
                       pcd_res;
                     }
                   ->
@@ -63,7 +63,12 @@ let init () =
                     let pcd_args =
                       match pcd_args with
                       | Pcstr_tuple pcd_args -> pcd_args
-                      | Pcstr_record _ -> assert false
+                      | Pcstr_record _ ->
+                        Location.raise_errorf ~loc:pcd_loc
+                          "@deriving(accessors) from a variant record argument \
+                           is unsupported. Either define the record type \
+                           separately from the variant type or use a \
+                           positional argument."
                     in
                     let little_con_name =
                       Ext_string.uncapitalize_ascii con_name
@@ -146,14 +151,19 @@ let init () =
                     {
                       pcd_name = {loc; txt = con_name};
                       pcd_args;
-                      pcd_loc = _;
+                      pcd_loc;
                       pcd_res;
                     }
                   ->
                     let pcd_args =
                       match pcd_args with
                       | Pcstr_tuple pcd_args -> pcd_args
-                      | Pcstr_record _ -> assert false
+                      | Pcstr_record _ ->
+                        Location.raise_errorf ~loc:pcd_loc
+                          "@deriving(accessors) from a variant record argument \
+                           is unsupported. Either define the record type \
+                           separately from the variant type or use a \
+                           positional argument."
                     in
                     let arity = pcd_args |> List.length in
                     let annotate_type =